+++ CHANGELOG.md
@@ -9,6 +9,7 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 **New fixes in 1.45:**
 
 * [mse275976: Unable to see completed badges](http://meta.stackexchange.com/q/275976) (credit: Patrick Hofman)
+* [mso318781: Code-only answers override the purple deleted colour](http://meta.stackoverflow.com/q/318781)
 
 **Other changes:**